backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
python-dotenv
pandas
numpy
scikit-learn
joblib
lightgbm
